add scopes for vm's
