+++ app/src/main/java/ru/tagilov/avitotrainee/city/di/CityModule.kt
@@ -10,29 +10,6 @@ import ru.tagilov.avitotrainee.city.ui.viewmodel.CityViewModel
 import ru.tagilov.avitotrainee.root.di.CityDependencies
 import javax.inject.Scope
 
-@Scope
-@kotlin.annotation.Retention(AnnotationRetention.RUNTIME)
-annotation class CityScreenScope
-
-@CityScreenScope
-@Component(
-        modules = [CityModule::class],
-        dependencies = [CityDependencies::class]
-)
-interface CityComponent {
-
-    @CityScreenScope
-    @FlowPreview
-    fun getViewModel(): CityViewModel
-
-    @Component.Builder
-    interface Builder {
-        fun deps(dependencies: CityDependencies): Builder
-        fun build(): CityComponent
-    }
-}
-
-
 @Module(includes = [CityBinds::class])
 class CityModule
