+++ app/src/main/java/ru/tagilov/avitotrainee/root/di/AppModule.kt
@@ -16,13 +16,18 @@ import ru.tagilov.avitotrainee.core.util.addQueryApiKey
 import ru.tagilov.avitotrainee.forecast.data.ForecastApi
 import ru.tagilov.avitotrainee.forecast.data.LocationApi
 import timber.log.Timber
+import java.lang.annotation.Documented
+import java.lang.annotation.RetentionPolicy
+import java.lang.annotation.RetentionPolicy.*
 import javax.inject.Qualifier
+import javax.inject.Scope
 import javax.inject.Singleton
+
 
 @Module
 class AppModule{
 
-    @Singleton //TODO поменять на нормальный скоуп
+    @AppScope //Есть вопросы
     @Provides
     fun provideDatabase(context: Context): AppDatabase {
         Timber.d("Database init!")
@@ -95,7 +100,6 @@ class AppModule{
             @Forecast forecastRetrofit: Retrofit
     ) : CityApi = forecastRetrofit.create()
 
-
 }
 
 
@@ -106,3 +110,6 @@ annotation class Location
 @Qualifier
 @Retention(AnnotationRetention.RUNTIME)
 annotation class Forecast
+
+@Scope
+annotation class AppScope
